# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Form Recognizer extracts information from forms and images into structured\
    \ data.  It includes the following options:\r\n* Form - Extracts information from\
    \ forms (PDFs and images) into structured data based on a model created from a\
    \ set of representative training forms. Form Recognizer learns the structure of\
    \ your forms to intelligently extract text and data. It ingests text from forms,\
    \ applies machine learning technology to identify keys, tables, and fields, and\
    \ then outputs structured data that includes the relationships within the original\
    \ file.\r\n* Receipt - Detects and extracts data from receipts using optical character\
    \ recognition (OCR) and our receipt model, enabling you to easily extract structured\
    \ data from receipts such as merchant name, merchant phone number, transaction\
    \ date, transaction total, and more.\r\n* Business Card - Detects and extracts\
    \ data from business cards using optical character recognition (OCR) and our business\
    \ card model, enabling you to easily extract structured data from business cards\
    \ such as contact names, company names, phone numbers, emails, and more.\r\n*\
    \ Layout - Extracts text and table structure from documents using optical character\
    \ recognition (OCR).\r\n* Invoices - Detects and extracts data from invoices using\
...
    \ models, enabling you to easily extract structured data from invoices such as\
...
host: "westus.api.cognitive.microsoft.com"
basePath: "/formrecognizer/v2.1-preview.3"
...
schemes:
...
security:
- apiKeyHeader: []
- apiKeyQuery: []
paths:
  /layout/analyze:
    post:
      summary: "Analyze Layout - Analyze Layout"
      description: "Extract text and layout information from a given document. The\
        \ input document must be of one of the supported content types - 'application/pdf',\
        \ 'image/jpeg', 'image/png' or 'image/tiff'. Alternatively, use 'application/json'\
        \ type to specify the Url location of the document to be analyzed.\r\n<br>\r\
        \nSupported languages: The Layout API is currently only available for Dutch,\
        \ English, French, German, Italian, Japanese, Portuguese, Simplified Chinese\
        \ and Spanish."
      operationId: "AnalyzeLayoutAsync"
      consumes:
      - "application/json"
      - "application/pdf"
      - "image/jpeg"
      - "image/png"
      - "image/tiff"
      produces:
      - "application/json"
      parameters:
      - name: "language"
        in: "query"
        description: "The BCP-47 language code of the text in the document. Currently,\
          \ only Afrikaans (‘af’), Albanian (‘sq’), Asturian (‘ast’), Basque (‘eu’\
          ), Bislama (‘bi’), Breton (‘br’), Catalan (‘ca’), Cebuano (‘ceb’), Chamorro\
          \ (‘ch’), Cornish (‘kw’), Corsican (‘co’), Crimean Tatar - Latin script(‘\
          crh’), Czech (‘cs’), Danish (‘da’), Dutch (‘nl’), English ('en'), Estonian\
          \ (‘et’), Fijian (‘fj’), Filipino (‘fil’), Finnish (‘fi’), French (‘fr’\
          ), Friulian (‘fur’), Galician (‘gl’), German (‘de’), Gilbertese (‘gil’),\
          \ Greenlandic (‘kl’), Haitian Creole (‘ht’), Hani (‘hni’), Hmong Daw (‘\
          mww’), Hungarian (‘hu’), Indonesian (‘id’), Interlingua (‘ia’), Inuktitut\
          \ (‘iu’), Irish (‘ga’), Italian (‘it’), Japanese (‘ja’), Javanese (‘jv’\
          ), Kabuverdianu (‘kea’), Kachin (‘kac’), Kara-Kalpak (‘kaa’), Kashubian\
          \ (‘csb’), Khasi (‘kha’), Korean (‘ko’), Kurdish - Latin script (‘ku’),\
          \ K’iche’ (‘quc’), Luxembourgish (‘lb’), Malay (‘ms’), Manx (‘gv’), Neapolitan\
          \ (‘nap’), Norwegian (‘no’), Occitan (‘oc’), Polish (‘pl’), Portuguese (‘\
          pt’), Romansh (‘rm’), Scots (‘sco’), Scottish Gaelic (‘gd’), simplified\
          \ Chinese (‘zh-Hans’), Slovenian (‘sl’), Spanish (‘es’), Swahili (‘sw’),\
          \ Swedish (‘sv’), Tatar - Latin script (‘tt’), Tetum (‘tet’), traditional\
          \ Chinese (‘zh-Hant’), Turkish (‘tr’), Upper Sorbian (‘hsb’), Uzbek (‘uz’\
          ), Volapük (‘vo’), Walser (‘wae’), Western Frisian (‘fy’), Yucatec Maya\
          \ (‘yua’), Zhuang (‘za’) and Zulu (‘zu’) are supported (print – seventy-three\
          \ languages and handwritten – English only). Layout supports auto language\
          \ identification and multilanguage documents, so only provide a language\
          \ code if you would like to force the documented to be processed as that\
          \ specific language."
        required: false
        type: "string"
        enum:
        - "af"
        - "ast"
        - "bi"
        - "br"
        - "ca"
        - "ceb"
        - "ch"
        - "co"
        - "crh"
        - "cs"
        - "csb"
        - "da"
        - "de"
        - "en"
        - "es"
        - "et"
        - "eu"
        - "fi"
        - "fil"
        - "fj"
        - "fr"
        - "fur"
        - "fy"
        - "ga"
        - "gd"
        - "gil"
        - "gl"
        - "gv"
        - "hni"
        - "hsb"
        - "ht"
        - "hu"
        - "ia"
        - "id"
        - "it"
        - "iu"
        - "ja"
        - "jv"
        - "kaa"
        - "kac"
        - "kea"
        - "kha"
        - "kl"
        - "ko"
        - "ku"
        - "kw"
        - "lb"
        - "ms"
        - "mww"
        - "nap"
        - "nl"
        - "no"
        - "oc"
        - "pl"
        - "pt"
        - "quc"
        - "rm"
        - "sco"
        - "sl"
        - "sq"
        - "sv"
        - "sw"
        - "tet"
        - "tr"
        - "tt"
        - "uz"
        - "vo"
        - "wae"
        - "yua"
        - "za"
        - "zh-Hans"
        - "zh-Hant"
        - "zu"
      - name: "pages"
        in: "query"
        description: "The page selection only leveraged for multi-page PDF and TIFF\
          \ documents. Accepted input include single pages (e.g.'1, 2' -> pages 1\
          \ and 2 will be processed), finite (e.g. '2-5' -> pages 2 to 5 will be processed)\
          \ and open-ended ranges (e.g. '5-' -> all the pages from page 5 will be\
          \ processed & e.g. '-10' -> pages 1 to 10 will be processed). All of these\
          \ can be mixed together and ranges are allowed to overlap (eg. '-5, 1, 3,\
          \ 5-10' - pages 1 to 10 will be processed). The service will accept the\
          \ request if it can process at least one page of the document (e.g. using\
          \ '5-100' on a 5 page document is a valid input where page 5 will be processed).\
          \ If no page range is provided, the entire document will be processed."
        required: false
        type: "string"
      - name: "readingOrder"
        in: "query"
        description: "Optional parameter to specify which reading order algorithm\
          \ should be applied when ordering the extract text elements. Can be either\
          \ 'basic' or 'natural'. Will default to basic if not specified"
        required: false
        type: "string"
        enum:
        - "basic"
        - "natural"
      - in: "body"
        name: "sourcePath"
        description: ".json, .pdf, .jpg, .png or .tiff type file stream."
        required: false
        schema:
          $ref: "#/definitions/SourcePath"
      responses:
        "202":
          description: "The service has accepted the request and will start processing\
            \ soon.  The client can query the operation status and result using the\
            \ URL specified in the 'Operation-Location' response header.  The URL\
            \ expires in 24 hours.\r\n<br/>\r\n<table class=\"element table\">\r\n\
            \    <thead/>\r\n    <tbody>\r\n        <tr>\r\n            <td>\r\n \
            \               <b>Operation-Location</b>\r\n            </td>\r\n   \
            \         <td>\r\n                URL containing the resultId used to\
            \ track the progress and obtain the result of the analyze operation.\r\
            \n                <br/>\r\n                Example: https://cognitiveservice/formrecognizer/v2.1-preview.3/layout/analyzeResults/{resultId}\r\
            \n            </td>\r\n        </tr>\r\n    </tbody>\r\n</table>\r\n<br/>\r\
            \nSample Response Header:\r\n<p><code>Content-Length: 0\r\nOperation-Location:\
            \ https://centraluseuap.api.cognitive.microsoft.com/formrecognizer/v2.1-preview.3/layout/analyzeResults/4ebd1503-3c14-4f90-9127-da0ecfec7bcd\r\
            \napim-request-id: 4ebd1503-3c14-4f90-9127-da0ecfec7bcd\r\nDate: Fri,\
            \ 04 Sep 2020 10:22:17 GMT\r\n</code></p>"
          headers:
            Operation-Location:
              type: "string"
              description: "URL containing the resultId used to track the progress\
                \ and obtain the result of the analyze operation."
          schema:
            $ref: "#/definitions/AnalyzeOperationResult"
        "400":
          description: "Bad request error. Detailed error code and message are specified\
            \ in the JSON response:\r\n<table class=\"element table\">\r\n    <thead>\r\
            \n        <tr>\r\n            <th>Error Code</th>\r\n            <th>Description</th>\r\
            \n        </tr>\r\n    </thead>\r\n    <tbody>\r\n        <tr>\r\n   \
            \         <td>BadArgument</td>\r\n            <td>Bad or unrecognizable\
            \ request JSON or binary file.</td>\r\n        </tr>\r\n        <tr>\r\
            \n            <td>InvalidImageURL</td>\r\n            <td>Image URL is\
            \ badly formatted.</td>\r\n        </tr>\r\n        <tr>\r\n         \
            \   <td>FailedToDownloadImage</td>\r\n            <td>Failed to download\
            \ image from input URL.</td>\r\n        </tr>\r\n        <tr>\r\n    \
            \        <td>InvalidImage</td>\r\n            <td>The input data is not\
            \ a valid image or password protected.</td>\r\n        </tr>\r\n     \
            \   <tr>\r\n            <td>UnsupportedImageFormat</td>\r\n          \
            \  <td>Image format unsupported. Supported formats include JPEG, PNG,\
            \ PDF and TIFF.</td>\r\n        </tr>\r\n        <tr>\r\n            <td>InvalidImageSize</td>\r\
            \n            <td>The input image is too large. It should not be larger\
            \ than 50MB.</td>\r\n        </tr>\r\n        <tr>\r\n            <td>InvalidImageDimension</td>\r\
            \n            <td>\r\n                The input image dimension is out\
            \ of range. The minimum image dimension is 50 x 50 pixels and the maximum\
            \ is 10000 x 10000 pixels. The maximum PDF dimension is 17 x 17 inches.\r\
            \n            </td>\r\n        <tr>\r\n            <td>InvalidReadingOrder</td>\r\
            \n            <td>\r\n                The specified Reading Order is invalid.\r\
            \n            </td>\r\n        </tr>\r\n        <tr>\r\n            <td>InvalidPageRange</td>\r\
            \n            <td>\r\n                The specified Pages parameter is\
            \ invalid. Please see returned error for more details.\r\n           \
            \ </td>\r\n        </tr>\r\n    </tbody>\r\n</table>"
          examples:
            application/json:
              error:
                code: "BadArgument"
                message: "Invalid input."
          schema:
            $ref: "#/definitions/ErrorResponse"
        "415":
          description: "Unsupported media type error. 'Content-Type' does not match\
            \ the POST content.\r\n<br/>\r\n<ul>\r\n    <li>For image URL, 'Content-Type'\
            \ should be 'application/json'.</li>\r\n    <li>For binary PDF data, 'Content-Type'\
            \ should be 'application/pdf'.</li>\r\n    <li>For binary image data,\
            \ 'Content-Type' should be 'image/jpeg', 'image/png' or 'image/tiff'.</li>\r\
            \n</ul>\r\n<br/>\r\n"
          examples:
            application/json:
              error:
                code: "BadArgument"
                message: "Unsupported media type."
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error."
          examples:
            application/json:
              error:
                code: "Unspecified"
                message: "Internal server error."
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: "Transient fault while querying Microsoft Azure storage services."
          examples:
            application/json:
              error:
                code: "StorageException"
                message: "Transient fault occurred while querying Microsoft Azure\
                  \ storage services. Please try again later."
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Default"
  /layout/analyzeResults/{resultId}:
    get:
      summary: "Analyze Layout - Get Analyze Layout Result"
      description: "Track the progress and obtain the result of the analyze layout\
        \ operation"
      operationId: "GetAnalyzeLayoutResult"
      produces:
      - "application/json"
      parameters:
      - name: "resultId"
        in: "path"
        description: "Format - uuid. Analyze operation result identifier."
        required: true
        type: "string"
      responses:
        "200":
          description: "JSON fields in the response body:\r\n<table class=\"element\
            \ table\">\r\n  <thead>\r\n    <tr>\r\n      <th>Field</th>\r\n      <th>Type</th>\r\
            \n      <th>Description</th>\r\n    </tr>\r\n  </thead>\r\n  <tbody>\r\
            \n    <tr>\r\n      <td>status</td>\r\n      <th>String</th>\r\n     \
            \ <td><a href=\"/docs/services/form-recognizer-api-v2-preview/operations/AnalyzeLayoutAsync\"\
            >Analyze Layout</a> operation status.\r\n        Possible values:<ul>\r\
            \n          <li><b>notStarted</b>: The analysis operation has not started.</li>\r\
            \n          <li><b>running</b>: The analysis operation is in progress.</li>\r\
            \n          <li><b>failed</b>: The analysis operation has failed.</li>\r\
            \n          <li><b>succeeded</b>: The analysis operation has succeeded.</li>\r\
            \n        </ul>If the status is <b>succeeded</b>, the response JSON will\
            \ include the layout and text recognition results. The layout result is\
            \ organized as a list of tables for each page. When includeTextDetails\
            \ is true, the detailed text recognition result is organized as a hierarchy\
            \ of lines and words, with text, bounding box and confidence information.</td>\r\
            \n    </tr>\r\n    <tr>\r\n      <td>createdDateTime</td>\r\n      <th>String</th>\r\
            \n      <td>Date and time (UTC) when the operation was created.</td>\r\
            \n    </tr>\r\n    <tr>\r\n      <td>lastUpdatedDateTime</td>\r\n    \
            \  <th>String</th>\r\n      <td>Date and time (UTC) when the status is\
            \ last updated.</td>\r\n    </tr>\r\n    <tr>\r\n      <td>analyzeResult</td>\r\
            \n      <th>Object</th>\r\n      <td>Results of the analyze operation.</td>\r\
            \n    </tr>\r\n    <tr>\r\n      <td>version</td>\r\n      <th>String</th>\r\
            \n      <td>The version of schema used for this result.</td>\r\n    </tr>\r\
            \n    <tr>\r\n      <td>readResults</td>\r\n      <th>[Object]</th>\r\n\
            \      <td>List of of extracted text result for each page in the input\
            \ document.</td>\r\n    </tr>\r\n    <tr>\r\n      <td>page</td>\r\n \
            \     <th>Integer</th>\r\n      <td>The 1-based page number in the input\
            \ document.</td>\r\n    </tr>\r\n    <tr>\r\n      <td>width</td>\r\n\
            \      <th>Number</th>\r\n      <td>The width of the image/PDF in pixels/inches,\
            \ respectively.</td>\r\n    </tr>\r\n    <tr>\r\n      <td>height</td>\r\
            \n      <th>Number</th>\r\n      <td>The height of the image/PDF in pixels/inches,\
            \ respectively.</td>\r\n    </tr>\r\n    <tr>\r\n      <td>unit</td>\r\
            \n      <th>String</th>\r\n      <td>The unit used by the width, height\
            \ and boundingBox properties. For images, the unit is \"pixel\". For PDF,\
            \ the unit is \"inch\".</td>\r\n    </tr>\r\n    <tr>\r\n      <td>angle</td>\r\
            \n      <th>Number</th>\r\n      <td>The general orientation of the text\
            \ in clockwise direction, measured in degrees between (-180, 180].</td>\r\
            \n    </tr>\r\n    <tr>\r\n      <td>language</td>\r\n      <th>String</th>\r\
            \n      <td>The detected language on the page overall or of a specific\
            \ line if different from the page.</td>\r\n    </tr>\r\n    <tr>\r\n \
            \     <td>lines</td>\r\n      <th>[Object]</th>\r\n      <td>\r\n    \
            \    List of text lines. The maximum number of lines returned is 300 per\
            \ page. The lines are sorted top to bottom, left to right, although in\
            \ certain cases proximity is treated with higher priority. As the sorting\
            \ order depends on the detected text, it may change across images and\
            \ OCR version updates. Thus, business logic should be built upon the actual\
            \ line location instead of order. This field only appears when <b>includeTextDetails</b>\
            \ is set to true.\r\n      </td>\r\n    </tr>\r\n    <tr>\r\n      <td>words</td>\r\
            \n      <th>[Object]</th>\r\n      <td>List of words in the text line.</td>\r\
            \n    </tr>\r\n    <tr>\r\n      <td>boundingBox</td>\r\n      <th>[Number]</th>\r\
            \n      <td>\r\n        Quadrangle bounding box of a line or word, depending\
            \ on the parent object, specified as a list of 8 numbers. The coordinates\
            \ are specified relative to the top-left of the original image. The eight\
            \ numbers represent the four points, clockwise from the top-left corner\
            \ relative to the text orientation. For image, the (x, y) coordinates\
            \ are measured in pixels. For PDF, the (x, y) coordinates are measured\
            \ in inches.\r\n      </td>\r\n    </tr>\r\n    <tr>\r\n      <td>text</td>\r\
            \n      <th>String</th>\r\n      <td>Extracted field text in documentResults,\
            \ or the text content of a word/line in readResults.</td>\r\n    </tr>\r\
            \n    <tr>\r\n      <td>confidence</td>\r\n      <th>Number</th>\r\n \
            \     <td>Confidence value.</td>\r\n    </tr>\r\n    <tr>\r\n      <td>pageResults</td>\r\
            \n      <th>[Object]</th>\r\n      <td>Page-level information extracted\
            \ from the input.</td>\r\n    </tr>\r\n    <tr>\r\n      <td>page</td>\r\
            \n      <th>Integer</th>\r\n      <td>Page number.</td>\r\n    </tr>\r\
            \n    <tr>\r\n      <td>tables</td>\r\n      <th>[Object]</th>\r\n   \
            \   <td>List of data tables extracted from the page.</td>\r\n    </tr>\r\
            \n    <tr>\r\n      <td>rows</td>\r\n      <th>Integer</th>\r\n      <td>Number\
            \ of rows.</td>\r\n    </tr>\r\n    <tr>\r\n      <td>columns</td>\r\n\
            \      <th>Integer</th>\r\n      <td>Number of columns.</td>\r\n    </tr>\r\
            \n    <tr>\r\n      <td>cells</td>\r\n      <th>[Object]</th>\r\n    \
            \  <td>List of cells contained in the table.</td>\r\n    </tr>\r\n   \
            \ <tr>\r\n      <td>rowIndex</td>\r\n      <th>Integer</th>\r\n      <td>Row\
            \ index of the cell.</td>\r\n    </tr>\r\n    <tr>\r\n      <td>columnIndex</td>\r\
            \n      <th>Integer</th>\r\n      <td>Column index of the cell.</td>\r\
            \n    </tr>\r\n    <tr>\r\n      <td>rowSpan</td>\r\n      <th>Integer</th>\r\
            \n      <td>Number of rows spanned by this cell.</td>\r\n    </tr>\r\n\
            \    <tr>\r\n      <td>columnSpan</td>\r\n      <th>Integer</th>\r\n \
            \     <td>Number of columns spanned by this cell.</td>\r\n    </tr>\r\n\
            \    <tr>\r\n      <td>elements</td>\r\n      <th>[String]</th>\r\n  \
            \    <td>When includeTextDetails is set to true, a list of JSON references\
            \ to the text elements constituting this field.</td>\r\n    </tr>\r\n\
            \    <tr>\r\n      <td>selectionMarks</td>\r\n      <th>[Object]</th>\r\
            \n      <td>List of detected selection marks</td>\r\n    </tr>\r\n   \
            \ <tr>\r\n      <td>state</td>\r\n      <th>String</th>\r\n      <td>State\
            \ of the selection mark. Can be either 'selected'/'unselected'</td>\r\n\
            \    </tr>\r\n    <tr>\r\n      <td>appearance</td>\r\n      <th>Object</th>\r\
            \n      <td>An object describing the style of the line along with the\
            \ qualitative confidence score.</td>\r\n    </tr>\r\n    <tr>\r\n    \
            \  <td>style</td>\r\n      <th>String</th>\r\n      <td>The general style\
            \ of the line of text. Possible values:\r\n        <ul>\r\n          <li><b>handwriting</b>:\
            \ handwritten styled text.</li>\r\n          <li><b>other</b>: other text\
            \ styles.</li>\r\n        </ul>\r\n      </td>\r\n    <tr>\r\n      <td>styleConfidence</td>\r\
            \n      <th>Number</th>\r\n      <td>Confidence value between 0 and 1\
            \ inclusive.</td>\r\n    </tr>\r\n  </tbody>\r\n</table>\r\n"
          schema:
            $ref: "#/definitions/AnalyzeOperationResult"
        "404":
          description: "Invalid or expired result identifier."
          examples:
            application/json:
              error:
                code: "BadArgument"
                message: "Operation ID is invalid or expired."
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error."
          examples:
            application/json:
              error:
                code: "Unspecified"
                message: "Internal server error."
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: "Transient fault while querying Microsoft Azure storage services."
          examples:
            application/json:
              error:
                code: "StorageException"
                message: "Transient fault occurred while querying Microsoft Azure\
                  \ storage services. Please try again later."
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Default"
  /prebuilt/receipt/analyze:
    post:
      summary: "Analyze Receipt - Analyze Receipt"
      description: "Extract field text and semantic values from a given receipt document.\r\
        \nThe input document must be of one of the supported content types - 'application/pdf',\
        \ 'image/jpeg', 'image/png' or 'image/tiff'. Alternatively, use 'application/json'\
        \ type to specify the Url location of the document to be analyzed.\r\n\r\n\
        Note: this technology is currently only available for English receipts, with\
        \ one receipt per page."
      operationId: "AnalyzeReceiptAsync"
      consumes:
      - "application/json"
      - "application/pdf"
      - "image/jpeg"
      - "image/png"
      - "image/tiff"
      produces:
      - "application/json"
      parameters:
      - name: "includeTextDetails"
        in: "query"
        description: "Include text lines and element references in the result.  Default:\
          \ false."
        required: false
        type: "boolean"
      - name: "locale"
        in: "query"
        description: "Locale of the receipt. Supported locales: en-AU, en-CA, en-GB,\
          \ en-IN, en-US."
        required: false
        type: "string"
        enum:
        - "en-AU"
        - "en-CA"
        - "en-GB"
        - "en-IN"
        - "en-US"
      - name: "pages"
        in: "query"
        description: "The page selection for multi-page PDF and TIFF documents, to\
          \ extract Receipt information from individual pages and a range of pages\
          \ (like page 2, and pages 5-7) by entering the page numbers and ranges separated\
          \ by commas (e.g. '2, 5-7'). If not set, all pages will  be processed."
        required: false
        type: "string"
      - in: "body"
        name: "sourcePath"
        description: "Document containing the receipt image(s) to be analyzed. The\
          \ POST body should be the raw image binary or the image URL\r\nin JSON.<br\
          \ />Additional requirements:<ul>\r\n  <li>Image format must be one of JPEG,\
          \ PNG, PDF or TIFF.</li>\r\n  <li>For PDF and TIFF, only the first 200 pages\
          \ are processed. <ul>\r\n      <li>For free tier subscribers, only the first\
          \ 2 pages are processed.</li>\r\n    </ul>\r\n  </li>\r\n  <li>Image file\
          \ size must be less than 50 MB.</li>\r\n  <li>Image dimensions must be at\
          \ least 50 x 50 pixels and at most 10000 x 10000 pixels. PDF dimensions\
          \ must be at most\r\n    17 x 17 inches, corresponding to Legal or A3 paper\
          \ sizes and smaller.</li>\r\n</ul>"
        required: false
        schema:
          $ref: "#/definitions/SourcePath"
      responses:
        "202":
          description: "The service has accepted the request and will start processing\
            \ soon.  The client can query the operation status and result using the\
            \ URL specified in the 'Operation-Location' response header.  The URL\
            \ expires in 48 hours.\r\n<br/>\r\n<table class=\"element table\">\r\n\
            \    <thead/>\r\n    <tbody>\r\n        <tr>\r\n            <td>\r\n \
            \               <b>Operation-Location</b>\r\n            </td>\r\n   \
            \         <td>\r\n                URL containing the resultId used to\
            \ track the progress and obtain the result of the analyze operation.\r\
            \n                <br/>\r\n                Example: https://cognitiveservice/formrecognizer/v2.1-preview.1/prebuilt/receipt/analyzeResults/{resultId}\r\
            \n            </td>\r\n        </tr>\r\n    </tbody>\r\n</table>"
          headers:
            Operation-Location:
              type: "string"
              description: "URL containing the resultId used to track the progress\
                \ and obtain the result of the analyze operation. This URL expires\
                \ in 48 hours."
          schema:
            $ref: "#/definitions/AnalyzeOperationResult"
        "400":
          description: "Bad request error. Detailed error code and message are specified\
            \ in the JSON response:\r\n<table class=\"element table\">\r\n    <thead>\r\
            \n        <tr>\r\n            <th>Error Code</th>\r\n            <th>Description</th>\r\
            \n        </tr>\r\n    </thead>\r\n    <tbody>\r\n        <tr>\r\n   \
            \         <td>BadArgument</td>\r\n            <td>Bad or unrecognizable\
            \ request JSON or binary file.</td>\r\n        </tr>\r\n        <tr>\r\
            \n            <td>FailedToDownloadImage</td>\r\n            <td>Failed\
            \ to download image from input URL.</td>\r\n        </tr>\r\n        <tr>\r\
            \n            <td>InvalidImage</td>\r\n            <td>The input data\
            \ is not a valid image or password protected.</td>\r\n        </tr>\r\n\
            \        <tr>\r\n            <td>InvalidImageDimension</td>\r\n      \
            \      <td>\r\n                The input image dimension is out of range.\
            \ The minimum image dimension is 50 x 50 pixels and the maximum is 10000\
            \ x 10000 pixels. The maximum PDF dimension is 17 x 17 inches.\r\n   \
            \         </td>\r\n        </tr>\r\n        <tr>\r\n            <td>InvalidImageSize</td>\r\
            \n            <td>The input image is too large. It should not be larger\
            \ than 50MB.</td>\r\n        </tr>\r\n        <tr>\r\n            <td>InvalidImageURL</td>\r\
            \n            <td>Image URL is badly formatted.</td>\r\n        </tr>\r\
            \n        <tr>\r\n            <td>UnsupportedImageFormat</td>\r\n    \
            \        <td>Image format unsupported. Supported formats include JPEG,\
            \ PNG, PDF and TIFF.</td>\r\n        </tr>\r\n        <tr>\r\n       \
            \     <td>UnsupportedLocale</td>\r\n            <td>Locale unsupported.\
            \ Supported locales include en-AU, en-CA, en-GB, en-IN and en-US.</td>\r\
            \n        </tr>\r\n    </tbody>\r\n</table>\r\n"
          examples:
            application/json:
              error:
                code: "BadArgument"
                message: "Invalid input."
          schema:
            $ref: "#/definitions/ErrorResponse"
        "415":
          description: "Unsupported media type error. 'Content-Type' does not match\
            \ the POST content.\r\n<br/>\r\n<ul>\r\n    <li>For image URL, 'Content-Type'\
            \ should be 'application/json'.</li>\r\n    <li>For binary PDF data, 'Content-Type'\
            \ should be 'application/pdf'.</li>\r\n    <li>For binary image data,\
            \ 'Content-Type' should be 'image/jpeg', 'image/png' or 'image/tiff'.</li>\r\
            \n</ul>\r\n<br/>\r\n"
          examples:
            application/json:
              error:
                code: "BadArgument"
                message: "Unsupported media type."
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error."
          examples:
            application/json:
              error:
                code: "Unspecified"
                message: "Internal server error."
          schema:
            $ref: "#/definitions/ErrorResponse"
        "503":
          description: "Transient fault while querying Microsoft Azure storage services.\r\
            \n"
          examples:
            application/json:
              error:
                code: "StorageException"
                message: "Transient fault occurred while querying Microsoft Azure\
                  \ storage services. Please try again later."
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "Default"
...
securityDefinitions:
  apiKeyHeader:
    type: "apiKey"
    name: "Ocp-Apim-Subscription-Key"
    in: "header"
  apiKeyQuery:
    type: "apiKey"
    name: "subscription-key"
    in: "query"
definitions:
...
  ErrorResponse:
    type: "object"
    required:
    - "error"
    properties:
      error:
        $ref: "#/definitions/ErrorInformation"
  ErrorInformation:
    type: "object"
    required:
    - "code"
    - "message"
    properties:
      code:
        type: "string"
      message:
        type: "string"
...
  SourcePath:
    type: "object"
    properties:
      source:
        type: "string"
        description: "Url path."
        minLength: 0
        maxLength: 2048
    description: "Url to source data."
    example: "{ \"source\": \"http://example.com/test.jpg\" }"
...